move camera to module folder
