+++ controller/modules/camera/api.go
