+++ controller/modules/camera/camera.go
