+++ controller/modules/camera/config.go
