+++ controller/modules/camera/process.go
